Updated code to be compatible with Python 3.  All examples and tests run correctly, except the params.use_tex(True) example.  The params.use_tex(True) issue may be due to my setup.

+++ doc/source/examples/video/video.py
@@ -66,7 +66,7 @@ def update_fig(i):
     
 #Create the animation
 anim = ani.FuncAnimation(fig.mpl_fig, update_fig, init_func = init, \
-    frames = max(ndx_list)/2, interval = 2)
+    frames = int(max(ndx_list)/2), interval = 2)
 #Specify ffmpeg installation path
 plt.rcParams['animation.ffmpeg_path'] = u'/opt/local/bin/ffmpeg'
 #Save the animation
